Agrega modificaciones finales, funcionamiento total

+++ main.py
@@ -15,22 +15,46 @@ usuarioActual = {"nombre":"", "esAdmin":False, "icon":""
                 }
 # SECCIÓN INICIO DE SESION
 
+@app.route('/juegoNuevo')
+def juegoNuevo():
+  preguntas1 = listaPreguntasSegunNivel('1')
+  preguntas2 = [
+  listaPreguntasSegunCategoria('ODS 4'),
+  listaPreguntasSegunCategoria('ODS 10'),
+  listaPreguntasSegunCategoria('ODS 15')
+  ]
+  preguntasBonus = listaPreguntasSegunNivel('bonus')
+  respuestas = listaRespuestas()
+  datos = datosUsuarioActual(usuarioActual["nombre"])
+  topUsuario=listaTopUsuarios()
+  return render_template("juego.html",usuario=usuarioActual["nombre"], preguntas1=preguntas1,
+                              preguntas2=preguntas2,
+                              preguntasBonus=preguntasBonus,
+                              respuestas=respuestas,
+                              datos=datos,
+                              topUsuario=topUsuario)
+  
 
 @app.route('/')
 def index():
-    return render_template('index.html')
+  global usuarioActual
+  usuarioActual = {"nombre":"", "esAdmin":False, "icon":"" 
+                }
+  return render_template('index.html')
 
 
 @app.route('/iniciarsesion')
 def iniciarSesion():
-    return render_template('inicio_sesion.html')
+    msg = 'todo bien'
+    return render_template('inicio_sesion.html', msg=msg)
 
 
 #https://lineadecodigo.com/python/sesion-en-flask/
 #https://www.geeksforgeeks.org/how-to-use-flask-session-in-python-flask/
 @app.route('/crearusuario')
 def crear():
-    return render_template("crear_usuario.html")
+    msg = 'todo bien'
+    return render_template("crear_usuario.html", msg=msg)
 
 
 @app.route('/ingresarusuario', methods=['POST', 'GET'])
@@ -46,7 +70,7 @@ def crearUsuario():
                 session['contraseña'] = contraseña
                 existe = checkearSiEsta(usuario)
                 if existe == True:
-                    msg = "Su usario ya existe. Por favor intentelo nuevamente"
+                    msg = "Su usario ya existe. Por favor intentelo nuevamente o inicie sesion"
                     return render_template("crear_usuario.html", msg=msg)
                 else:
                     msg = f"{usuario}"
@@ -63,17 +87,18 @@ def crearUsuario():
                     preguntasBonus = listaPreguntasSegunNivel('bonus')
                     respuestas = listaRespuestas()
                     datos = datosUsuarioActual(usuario)
+                    topUsuario=listaTopUsuarios()
+                  
                     return render_template("juego.html",usuario=usuario, preguntas1=preguntas1,
-                               preguntas2=preguntas2,
-                               preguntasBonus=preguntasBonus,
-                               respuestas=respuestas,
-                                datos=datos)# ACA ESTA EL PROBLEMA
+                              preguntas2=preguntas2,
+                              preguntasBonus=preguntasBonus,
+                              respuestas=respuestas,
+                              datos=datos,
+                              topUsuario=topUsuario)
             else:
                 msg = "Tu contraseña no ha sido confirmada. Intentalo nuevamente"
                 return render_template("crear_usuario.html", msg=msg)
-        else:
-            msg = "Debe ingresar usuario, contraseña y confirmacion. Intentelo nuevamente"
-            return render_template("crear_usuario.html", msg=msg)
+        return render_template("crear_usuario.html", msg=msg)
 
 
 #Checkea que el usuario no exista
@@ -119,15 +144,15 @@ def logearte():
         preguntasBonus = listaPreguntasSegunNivel('bonus')
         respuestas = listaRespuestas()
         datos = datosUsuarioActual(usuario)
-        #topUsuario=listaTopUsuarios()
+        topUsuario=listaTopUsuarios()
         return render_template('juego.html',
                                usuario=usuario,
                                preguntas1=preguntas1,
                                preguntas2=preguntas2,
                                preguntasBonus=preguntasBonus,
                                respuestas=respuestas,
-                                datos=datos)
-                                #topUsuario=topUsuario)
+                                datos=datos,
+                              topUsuario=topUsuario)
       
     else:
         msg = 'Usuario inexistente. Intentelo nuevamente'
@@ -137,13 +162,14 @@ def logearte():
 @app.route('/borrar/<unUsuario>')
 def borrarUsuario(unUsuario):
   conn = sqlite3.connect('dataBase.db')
-  cur = conn.cursor()
-  cur.execute(f"""DELETE FROM Usuarios
+  if usuarioActual["esAdmin"]:
+    cur = conn.cursor()
+    cur.execute(f"""DELETE FROM Usuarios
                   WHERE nombre = '{unUsuario}';
                   """)
-  conn.commit()
+    conn.commit()
   conn.close()
-  return "funcionooooo"
+  return "Usuario Borrado"
 
 
 #Modifica variable global en cuanto al ingreso de sesion
@@ -172,7 +198,7 @@ def userowo(n):
 # Ingresar usuario a base de datos
 def insertUser(unUsuario, unaContraseña, unAvatar):
     con = sqlite3.connect("dataBase.db")
-    q = f"""INSERT INTO Usuarios (nombre,contraseña,icon) VALUES ('{unUsuario}', '{unaContraseña}', '{unAvatar}')"""
+    q = f"""INSERT INTO Usuarios (nombre,contraseña,icon, mayor_puntaje) VALUES ('{unUsuario}', '{unaContraseña}', '{unAvatar}', 0)"""
     con.execute(q)
     con.commit()
     con.close()
@@ -207,6 +233,7 @@ def listaPreguntas():
 ## Lista todas las preguntas
 @app.route('/listaUsuarios')
 def listaUsuarios():
+  if usuarioActual["esAdmin"] == True:
     conn = sqlite3.connect('dataBase.db')
     conn.row_factory = sqlite3.Row
     cur = conn.cursor()
@@ -226,6 +253,8 @@ def listaUsuarios():
         }
         usuarios.append(usuario)
     return jsonify(usuarios)
+  else:
+    return render_template('noAdmin.html') 
 
 
 # Lista las preguntas según Nivel
@@ -255,8 +284,10 @@ def listaPreguntasSegunNivel(nivel):
         if preguntas[numero_random] not in preguntas_random:
             preguntas_random.append(preguntas[numero_random])
 
-
-    return preguntas_random
+    if usuarioActual["esAdmin"]:
+      return preguntas_random
+    else:
+      return render_template('noAdmin.html')
 
 
     # Lista las preguntas según Nivel
@@ -312,58 +343,58 @@ def listaRespuestas():
 
 
 ## Muestra la lista con las categorias posibles
-@app.route('/listaCategorias')
-def listaCategorias():
-    conn = sqlite3.connect('dataBase.db')
-    conn.row_factory = sqlite3.Row
-    cur = conn.cursor()
-    cur.execute(f"""SELECT categoria
-            FROM Preguntas
-            GROUP BY categoria
-                    """)
-    rows = cur.fetchall()
-    categorias = []
-    for fila in rows:
-        categorias.append(fila[0])
-    return jsonify(categorias)
+# @app.route('/listaCategorias')
+# def listaCategorias():
+#     conn = sqlite3.connect('dataBase.db')
+#     conn.row_factory = sqlite3.Row
+#     cur = conn.cursor()
+#     cur.execute(f"""SELECT categoria
+#             FROM Preguntas
+#             GROUP BY categoria
+#                     """)
+#     rows = cur.fetchall()
+#     categorias = []
+#     for fila in rows:
+#         categorias.append(fila[0])
+#     return jsonify(categorias)
 
 
-## Muestra la lista con los nombres de los niveles
-@app.route('/listaNiveles')
-def listaNiveles():
-    conn = sqlite3.connect('dataBase.db')
-    conn.row_factory = sqlite3.Row
-    cur = conn.cursor()
-    cur.execute(f"""SELECT categoria
-            FROM Preguntas
-            GROUP BY nivel
-                    """)
-    rows = cur.fetchall()
-    niveles = []
-    for fila in rows:
-        niveles.append(fila[0])
-    return jsonify(niveles)
+# ## Muestra la lista con los nombres de los niveles
+# @app.route('/listaNiveles')
+# def listaNiveles():
+#     conn = sqlite3.connect('dataBase.db')
+#     conn.row_factory = sqlite3.Row
+#     cur = conn.cursor()
+#     cur.execute(f"""SELECT categoria
+#             FROM Preguntas
+#             GROUP BY nivel
+#                     """)
+#     rows = cur.fetchall()
+#     niveles = []
+#     for fila in rows:
+#         niveles.append(fila[0])
+#     return jsonify(niveles)
   
 ## Mostrar lista top diez usuarios
-## @app.route('/listaTop10Usuarios')
+@app.route('/listaTop10Usuarios')
 def listaTopUsuarios():
     conn = sqlite3.connect('dataBase.db')
     conn.row_factory = sqlite3.Row
     cur = conn.cursor()
-    cur.execute(f"""SELECT nombre
+    cur.execute(f"""SELECT nombre, mayor_puntaje
             FROM Usuarios
+            WHERE NOT mayor_puntaje = 0
             ORDER BY mayor_puntaje desc limit 10
                     """)
     rows = cur.fetchall()
     topUsuario = []
     for fila in rows:
-        topUsuario.append(fila[0])
+        userFila = {}
+        userFila['nombre']=fila[0]
+        userFila['puntos']=fila[1]
+        topUsuario.append(userFila)
     return topUsuario
 
-#Editar preguntas/respuestas o Borrar Pregunta
-## Mostrar lista de puntaje
- ## @app.route('/listaPuntaje') 
-"""def listaPuntaje():"""
 
 @app.route('/editar', methods=['POST', 'GET'])
 def editaPregunta():
@@ -414,97 +445,31 @@ def adminEditar():
         return render_template('noAdmin.html') #HACER TEMPLATE
 
 
-# def crearPreg():
-  
-#   conn = sqlite3.connect('dataBase.db')
-#   preg = request.form['textoPregunta']
-#   nivel = request.form['nivel']
-#   categ = request.form['categoria']
-#   # if request.form.get('tipoPreguntaChoice'):
-#   #   tipo = request.form['tipoPreguntaChoice']
-#   # elif request.form.get('tipoPreguntaVoF'):
-#   #   tipo = request.form['tipoPreguntaVoF']
-#   a = request.form['tipoPregunta']
-#   print(a)
-#   resps = []
-#   listaCorrectas = []
-#   if request.form.get('correcto1'):
-#     listaCorrectas.append(1)
-#   elif request.form.get('correcto2'):
-#     listaCorrectas.append(1)
-#   elif request.form.get('correcto3'):
-#     listaCorrectas.append(1)
-    
-#   resps.append(request.form['rta1'])
-#   resps.append(request.form['rta2'])
-
-
-#   q = f"""INSERT INTO Preguntas(pregunta, categoria,nivel)
-#       VALUES('{preg}','{categ}','{nivel}');"""
-#   conn.execute(q)
-#   n = f"""SELECT id_pregunta
-#     FROM Preguntas
-#     ORDER BY id_pregunta DESC
-#     LIMIT 1"""
-#   conn.execute(n)
-#   cur = conn.cursor()
-#   id = cur.fetchall()
-#   i=0
-#   for resp in resps:
-#     conn.execute(f"""INSERT INTO Respuestas(es_correcta,respuesta,id_pregunta)
-#     VALUES({listaCorrectas[i]},{resp[i]},{id});""")
-#     i =i+1
-#   conn.commit()
-#   conn.close()
-  
-@app.route('/test', methods=['POST', 'GET'])
-def testing():
-  conn = sqlite3.connect('dataBase.db')
-  cur = conn.cursor()
-  id=3
-  cur.execute(f"""SELECT id_respuesta
-    FROM Respuestas
-    WHERE id_pregunta = {id};""")
-  ids = cur.fetchall()
-  print(ids)
-  i=0
-  for id in ids:
-    print(ids[i][0])
-    conn.execute(f"""UPDATE Respuestas
-    SET (respuesta)
-    VALUES respuesta = '{vector[i]}'
-    WHERE id_respuestas = '{ids[i][0]}'""")
-    i=i+1
-  return "mira vos"
-
 @app.route('/ingresarPuntaje', methods=['POST', 'GET'])
 def ingresarPuntaje():
   conn = sqlite3.connect('dataBase.db')
-  puntaje = "lol"
+  puntaje = request.form["puntajeFinal"]
   cur = conn.cursor()
   cur.execute(f"""SELECT mayor_puntaje
                     FROM Usuarios
                     WHERE nombre = '{usuarioActual['nombre']}';
                   """)
   user = cur.fetchall()
-  if user[0][0] == [] or user[0][0]<puntaje:
-    q = f"""INSERT INTO Usuarios (mayor_puntaje) VALUES ({puntaje})"""
+  print(user)
+  if user[0][0] == 0 or user[0][0]<int(puntaje):
+    q = f"""UPDATE Usuarios SET mayor_puntaje={puntaje} WHERE nombre = '{usuarioActual["nombre"]}'"""
     conn.execute(q)
     conn.commit()
   conn.close()
-  return ('ol')
+  return "eDUARDOOOOO VENI A BUSCAR A JUAN CRUZ"
     
 
-@app.route('/prueba', methods=['POST', 'GET'])
-def prueba():
-  # conn = sqlite3.connect('dataBase.db')
-  # puntaje = "lol"
-  # cur = conn.cursor()
-  # cur.execute(f"""SELECT mayor_puntaje
-  #                   FROM Usuarios
-  #                   WHERE nombre = '{usuarioActual['nombre']}';
-  #                 """)
-  # user = cur.fetchall()
-  # print(user)
+@app.route('/prueba/<n>', methods=['POST', 'GET'])
+def prueba(n):
+  conn = sqlite3.connect('dataBase.db')
+  q = f"""UPDATE Usuarios SET mayor_puntaje=0 WHERE nombre = '{n}'"""
+  conn.execute(q)
+  conn.commit()
+  conn.close()
   return "a"
 app.run(host='0.0.0.0', port=81)
